project_sam/requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
peft
pillow
opencv-python
sentencepiece
protobuf